[MIN-27] Removed pallet accounts
