+++ Cargo.toml
@@ -32,7 +32,6 @@ members = [
     'pallets/minterest-model',
     'pallets/minterest-protocol',
     'pallets/controller',
-    'pallets/accounts',
     'pallets/prices',
     'pallets/risk-manager',
     'pallets/liquidation-pools',
