--- pallets/accounts/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "accounts"
-description = 'Pallet for managing accounts and access rights in the project'
-version = "0.6.2"
-authors = ['Minterest <https://github.com/minterest-defi>']
-edition = "2018"
-
-[dependencies]
-serde = { version = "1.0.101", optional = true }
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
-frame-support = { version = '3.0.0', default-features = false }
-frame-system = { version = '3.0.0', default-features = false }
-sp-runtime = { version = '3.0.0', default-features = false }
-sp-std = { version = '3.0.0', default-features = false }
-
-[dev-dependencies]
-sp-core = { version = '3.0.0' }
-sp-io = { version = '3.0.0' }
-
-[features]
-default = ['std']
-std = [
-    "serde",
-    "codec/std",
-    'frame-support/std',
-    'frame-system/std',
-    'sp-runtime/std',
-    "sp-std/std",
-]
