--- pallets/accounts/rpc/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "accounts-rpc"
-version = "0.6.1"
-authors = ["Minterest <https://github.com/minterest-defi>"]
-edition = "2018"
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0" }
-jsonrpc-core = "15.0.0"
-jsonrpc-core-client = "15.0.0"
-jsonrpc-derive = "15.0.0"
-
-sp-runtime = { version = "3.0.0" }
-sp-api = { version = "3.0.0" }
-sp-blockchain = { version = "3.0.0" }
-
-accounts-rpc-runtime-api = { path = "runtime-api" }
