--- pallets/accounts/rpc/runtime-api/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "accounts-rpc-runtime-api"
-version = "0.6.1"
-authors = ["Minterest <https://github.com/minterest-defi>"]
-edition = "2018"
-license = "GPL-3.0-or-later"
-
-[dependencies]
-codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false, features = ["derive"] }
-sp-api = { version = "3.0.0", default-features = false }
-sp-std = { version = "3.0.0", default-features = false }
-sp-core = { version = "3.0.0", default-features = false }
-
-[dev-dependencies]
-serde_json = "1.0"
-
-[features]
-default = ["std"]
-std = [
-    "codec/std",
-    "sp-api/std",
-    "sp-std/std",
-]
