--- pallets/accounts/rpc/runtime-api/src/lib.rs
@@ -1,19 +1,0 @@
-//! Runtime API definition for the accounts pallet.
-
-#![cfg_attr(not(feature = "std"), no_std)]
-// The `too_many_arguments` warning originates from `decl_runtime_apis` macro.
-#![allow(clippy::too_many_arguments)]
-// The `unnecessary_mut_passed` warning originates from `decl_runtime_apis` macro.
-#![allow(clippy::unnecessary_mut_passed)]
-
-use codec::Codec;
-
-// Here we declare the runtime API. It is implemented it the `impl` block in
-// runtime amalgamator file (the `runtime/src/lib.rs`)
-sp_api::decl_runtime_apis! {
-	pub trait AccountsApi<AccountId> where
-		AccountId: Codec,
-	{
-		fn is_admin(caller: AccountId) -> Option<bool>;
-	}
-}
