+++ pallets/controller/rpc/runtime-api/src/lib.rs
@@ -6,7 +6,7 @@
 // The `unnecessary_mut_passed` warning originates from `decl_runtime_apis` macro.
 #![allow(clippy::unnecessary_mut_passed)]
 
-use codec::{Decode, Encode};
+use codec::{Codec, Decode, Encode};
 use minterest_primitives::{CurrencyId, Rate};
 use sp_core::RuntimeDebug;
 use sp_std::prelude::*;
@@ -25,7 +25,8 @@ pub struct PoolState {
 // Here we declare the runtime API. It is implemented it the `impl` block in
 // runtime amalgamator file (the `runtime/src/lib.rs`)
 sp_api::decl_runtime_apis! {
-	pub trait ControllerApi {
+	pub trait ControllerApi<AccountId> where AccountId: Codec {
 		fn liquidity_pool_state(pool_id: CurrencyId) -> Option<PoolState>;
+		fn is_admin(caller: AccountId) -> Option<bool>;
 	}
 }
