+++ pallets/controller/rpc/src/lib.rs
@@ -1,5 +1,6 @@
 //! RPC interface for the controller pallet.
 
+use codec::Codec;
 pub use controller_rpc_runtime_api::{ControllerApi as ControllerRuntimeApi, PoolState};
 use jsonrpc_core::{Error as RpcError, ErrorCode, Result};
 use jsonrpc_derive::rpc;
@@ -10,9 +11,12 @@ use sp_runtime::{generic::BlockId, traits::Block as BlockT};
 use std::sync::Arc;
 
 #[rpc]
-pub trait ControllerApi<BlockHash> {
+pub trait ControllerApi<BlockHash, AccountId> {
 	#[rpc(name = "controller_liquidityPoolState")]
 	fn liquidity_pool_state(&self, pool_id: CurrencyId, at: Option<BlockHash>) -> Result<Option<PoolState>>;
+
+	#[rpc(name = "controller_isAdmin")]
+	fn is_admin(&self, caller: AccountId, at: Option<BlockHash>) -> Result<Option<bool>>;
 }
 
 /// A struct that implements the [`ControllerApi`].
@@ -43,11 +47,12 @@ impl From<Error> for i64 {
 	}
 }
 
-impl<C, Block> ControllerApi<<Block as BlockT>::Hash> for Controller<C, Block>
+impl<C, Block, AccountId> ControllerApi<<Block as BlockT>::Hash, AccountId> for Controller<C, Block>
 where
 	Block: BlockT,
 	C: Send + Sync + 'static + ProvideRuntimeApi<Block> + HeaderBackend<Block>,
-	C::Api: ControllerRuntimeApi<Block>,
+	C::Api: ControllerRuntimeApi<Block, AccountId>,
+	AccountId: Codec,
 {
 	fn liquidity_pool_state(
 		&self,
@@ -61,6 +66,18 @@ where
 		api.liquidity_pool_state(&at, pool_id).map_err(|e| RpcError {
 			code: ErrorCode::ServerError(Error::RuntimeError.into()),
 			message: "Unable to get pool state.".into(),
+			data: Some(format!("{:?}", e).into()),
+		})
+	}
+
+	fn is_admin(&self, caller: AccountId, at: Option<<Block as BlockT>::Hash>) -> Result<Option<bool>> {
+		let api = self.client.runtime_api();
+		let at = BlockId::hash(at.unwrap_or_else(||
+			// If the block hash is not supplied assume the best block.
+			self.client.info().best_hash));
+		api.is_admin(&at, caller).map_err(|e| RpcError {
+			code: ErrorCode::ServerError(Error::RuntimeError.into()),
+			message: "Unable to check if is an admin.".into(),
 			data: Some(format!("{:?}", e).into()),
 		})
 	}
