+++ pallets/controller/src/lib.rs
@@ -77,9 +77,7 @@ pub mod module {
 	use super::*;
 
 	#[pallet::config]
-	pub trait Config:
-		frame_system::Config + liquidity_pools::Config + accounts::Config + minterest_model::Config
-	{
+	pub trait Config: frame_system::Config + liquidity_pools::Config + minterest_model::Config {
 		/// The overarching event type.
 		type Event: From<Event> + IsType<<Self as frame_system::Config>::Event>;
