+++ pallets/liquidation-pools/Cargo.toml
@@ -18,7 +18,6 @@ orml-traits = { path = "../../orml/traits", default-features = false }
 orml-utilities = { path = "../../orml/utilities", default-features = false }
 
 liquidity-pools = { path = "../liquidity-pools", default-features = false }
-accounts = { path = "../accounts", default-features = false }
 minterest-primitives = { path = "../../primitives", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
 
@@ -29,7 +28,6 @@ orml-currencies = { path = "../../orml/currencies" }
 orml-tokens = { path = "../../orml/tokens" }
 
 liquidity-pools = { path = "../liquidity-pools" }
-accounts = { path = "../accounts" }
 
 [features]
 default = ['std']
