+++ pallets/minterest-model/src/mock.rs
@@ -9,7 +9,6 @@ use orml_traits::parameter_type_with_key;
 use pallet_traits::PriceProvider;
 use sp_core::H256;
 use sp_runtime::traits::AccountIdConversion;
-use sp_runtime::traits::One;
 use sp_runtime::{
 	testing::Header,
 	traits::{BlakeTwo256, IdentityLookup},
@@ -29,7 +28,6 @@ frame_support::construct_runtime!(
 		System: frame_system::{Module, Call, Config, Storage, Event<T>},
 		Tokens: orml_tokens::{Module, Storage, Call, Event<T>, Config<T>},
 		TestMinterestModel: minterest_model::{Module, Storage, Call, Event, Config},
-		TestAccounts: accounts::{Module, Storage, Call, Event<T>, Config<T>},
 		TestPools: liquidity_pools::{Module, Storage, Call, Config<T>},
 	}
 );
@@ -82,15 +80,6 @@ impl orml_tokens::Config for Test {
 	type WeightInfo = ();
 	type ExistentialDeposits = ExistentialDeposits;
 	type OnDust = ();
-}
-
-parameter_types! {
-	pub const MaxMembers: u8 = MAX_MEMBERS;
-}
-
-impl accounts::Config for Test {
-	type Event = Event;
-	type MaxMembers = MaxMembers;
 }
 
 parameter_types! {
@@ -149,7 +138,6 @@ impl minterest_model::Config for Test {
 }
 
 pub const BLOCKS_PER_YEAR: u128 = 5_256_000;
-pub const MAX_MEMBERS: u8 = 16;
 pub const ALICE: AccountId = 1;
 pub fn alice() -> Origin {
 	Origin::signed(ALICE)
@@ -194,13 +182,6 @@ pub fn new_test_ext() -> sp_io::TestExternalities {
 		],
 	}
 	.assimilate_storage::<Test>(&mut t)
-	.unwrap();
-
-	accounts::GenesisConfig::<Test> {
-		allowed_accounts: vec![(ALICE, ())],
-		member_count: u8::one(),
-	}
-	.assimilate_storage(&mut t)
 	.unwrap();
 
 	let mut ext: sp_io::TestExternalities = t.into();
