+++ pallets/risk-manager/Cargo.toml
@@ -26,7 +26,6 @@ minterest-protocol = { path = "../minterest-protocol", default-features = false 
 liquidation-pools = { path = "../liquidation-pools", default-features = false }
 pallet-traits = { path = "../traits", default-features = false }
 controller = { path = "../controller", default-features = false }
-accounts = { path = "../accounts", default-features = false }
 
 [dev-dependencies]
 sp-core = { version = '3.0.0' }
@@ -37,7 +36,6 @@ orml-tokens = { path = "../../orml/tokens" }
 liquidity-pools = { path = "../liquidity-pools" }
 liquidation-pools = { path = "../liquidation-pools" }
 controller = { path = "../controller" }
-accounts = { path = "../accounts" }
 minterest-model = { path = "../minterest-model" }
 
 [features]
@@ -60,5 +58,4 @@ std = [
     "minterest-protocol/std",
     "liquidation-pools/std",
     "controller/std",
-    "accounts/std",
 ]
