+++ rpc/src/lib.rs
@@ -27,12 +27,10 @@ where
 	C::Api: pallet_transaction_payment_rpc::TransactionPaymentRuntimeApi<Block, Balance>,
 	C::Api:
 		orml_oracle_rpc::OracleRuntimeApi<Block, DataProviderId, CurrencyId, node_minterest_runtime::TimeStampedPrice>,
-	C::Api: controller_rpc::ControllerRuntimeApi<Block>,
-	C::Api: accounts_rpc::AccountsRuntimeApi<Block, AccountId>,
+	C::Api: controller_rpc::ControllerRuntimeApi<Block, AccountId>,
 	C::Api: BlockBuilder<Block>,
 	P: TransactionPool + 'static,
 {
-	use accounts_rpc::{Accounts, AccountsApi};
 	use controller_rpc::{Controller, ControllerApi};
 	use orml_oracle_rpc::{Oracle, OracleApi};
 	use pallet_transaction_payment_rpc::{TransactionPayment, TransactionPaymentApi};
@@ -58,8 +56,6 @@ where
 	io.extend_with(ControllerApi::to_delegate(Controller::new(client.clone())));
 
 	io.extend_with(OracleApi::to_delegate(Oracle::new(client.clone())));
-
-	io.extend_with(AccountsApi::to_delegate(Accounts::new(client)));
 
 	// Extend this RPC with a custom API by using the following syntax.
 	// `YourRpcStruct` should have a reference to a client, which is needed
